+++ CHANGELOG.md
@@ -33,6 +33,31 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unrealeased
 
+### New Features
+
+* VolumeModel: to_geo (generate .geo file), to_msh (generate .msh file)
+
+
+### Fixed
+
+* 
+
+
+### Performance improvements
+
+*
+
+
+### Refactorings
+
+* 
+* 
+* 
+
+
+
+## Unrealeased
+
 
 ### New Features